feat: move logic to util
