+++ packages/blocks/src/page-block/utils/cursor.ts
@@ -1,0 +1,24 @@
+import {
+  caretRangeFromPoint,
+  resetNativeSelection,
+  SelectionEvent,
+} from '../../__internal__';
+
+export function repairerContextMenuRange(e: SelectionEvent) {
+  const currentRange = window.getSelection()?.getRangeAt(0);
+  const pointRange = caretRangeFromPoint(e.x, e.y);
+  // repair browser context menu change selection can not go through blocks
+  if (
+    currentRange &&
+    pointRange &&
+    currentRange.isPointInRange(
+      pointRange.startContainer,
+      pointRange.startOffset
+    ) &&
+    currentRange.isPointInRange(pointRange.endContainer, pointRange.endOffset)
+  ) {
+    requestAnimationFrame(() => {
+      resetNativeSelection(currentRange);
+    });
+  }
+}
